// hd_frondend/src/app/modules/encuesta/project/model/Encuesta.ts
export class Encuesta {
    idEncuesta: number;
    idTicket: number;
    idUsuario: number;
    puntuacion: number;
    comentarios: string;
    nombreAgente: string;
    departamentoANombre: string;
    descripcion: string;
    estadoTicketNombre: string;
    fechaCreacion: string;
    idPrioridad: number;
    idSla: string;
    nombreUsuario: string;
    slaNombre: string;
    titulo: string;
}